feat: Add insights v2 enricher
