+++ go.mod
@@ -3,8 +3,8 @@ module github.com/safedep/vet
 go 1.23.2
 
 require (
-	buf.build/gen/go/safedep/api/grpc/go v1.5.1-20241023094826-0cc88affe8f4.1
-	buf.build/gen/go/safedep/api/protocolbuffers/go v1.35.1-20241023094826-0cc88affe8f4.1
+	buf.build/gen/go/safedep/api/grpc/go v1.5.1-20241108135208-96765ea7d4c0.1
+	buf.build/gen/go/safedep/api/protocolbuffers/go v1.35.1-20241108135208-96765ea7d4c0.1
 	github.com/AlecAivazis/survey/v2 v2.3.7
 	github.com/CycloneDX/cyclonedx-go v0.9.1
 	github.com/anchore/syft v1.14.1
